refactor: login 로직을 Button에서 제거
